a basic working version of the key GUI components
