--- jsonAPI/src/main/java/com/jsonAPI/AisDecodeMessageStore/App.java
@@ -1,19 +1,0 @@
-package com.jsonAPI.AisDecodeMessageStore;
-
-import java.io.IOException;
-
-import dk.tbsalling.aismessages.demo.SimpleDemoApp;
-
-/**
- * Hello world!
- */
-public class App {
-  public static void main(String[] args) {
-    System.out.println("Hello World!");
-    try {
-      SimpleDemoApp.main(args);
-    } catch (IOException e) {
-      e.printStackTrace();
-    }
-  }
-}
