--- jsonAPI/src/main/java/org/apache/struts/helloworld/action/HelloWorldAction.java
@@ -1,39 +1,0 @@
-package org.apache.struts.helloworld.action;
-
-import com.opensymphony.xwork2.ActionSupport;
-
-import org.apache.struts.helloworld.model.MessageStore;
-
-/**
- * Acts as a Struts 2 controller that responds to a user action by setting the value of the Message
- * model class, and returns a String result.
- *
- * @author Bruce Phillips
- */
-public class HelloWorldAction extends ActionSupport {
-
-  private static final long serialVersionUID = 1L;
-
-  /**
-   * The model class that stores the message to display in the view.
-   */
-  private MessageStore messageStore;
-
-  /*
-   * Creates the MessageStore model object and
-   * returns success.  The MessageStore model
-   * object will be available to the view.
-   * (non-Javadoc)
-   * @see com.opensymphony.xwork2.ActionSupport#execute()
-   */
-  public String execute() {
-    messageStore = new MessageStore();
-
-    return SUCCESS;
-  }
-
-  public MessageStore getMessageStore() {
-    return messageStore;
-  }
-
-}
