+++ jsonAPI/src/main/resources/struts.xml
@@ -7,29 +7,16 @@
 
     <constant name="struts.devMode" value="true"/>
 
-    <package name="basicstruts2" extends="struts-default">
+    <package name="ais-decode-message-store" extends="struts-default">
 
-        <!-- If no class attribute is specified the framework will assume success and 
-        render the result index.jsp -->
-        <!-- If no name value for the result node is specified the success value is the default -->
         <action name="index">
             <result>/index.jsp</result>
         </action>
 
-        <!-- If the URL is hello.action then call the execute method of class HelloWorldAction.
-        If the result returned by the execute method is success render the HelloWorld.jsp -->
-        <action name="hello" class="org.apache.struts.helloworld.action.HelloWorldAction"
-                method="execute">
-            <result name="success">/HelloWorld.jsp</result>
+        <action name="map" class="com.jsonAPI.AisDecodeMessageStore.action.MapAction">
+            <result name="success">/map.jsp</result>
         </action>
 
-        <!--TODO Rename or remove -->
-        <!-- If the URL is test.action then call the execute method of class HelloWorldAction.
-         If the result returned by the execute method is success render the HelloWorld.jsp -->
-        <action name="test" class="org.apache.struts.helloworld.action.TestDisplayAction"
-                method="execute">
-            <result name="success">/TestDisplay.jsp</result>
-        </action>
 
     </package>
