+++ jsonAPI/src/main/webapp/WEB-INF/web.xml
@@ -2,11 +2,10 @@
 <web-app xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" id="WebApp_ID" version="2.4"
          xmlns="http://java.sun.com/xml/ns/j2ee"
          xsi:schemaLocation="http://java.sun.com/xml/ns/j2ee http://java.sun.com/xml/ns/j2ee/web-app_2_4.xsd">
-    <display-name>Hello World Struts 2</display-name>
+    <display-name>AIS Decode Message Store</display-name>
     <welcome-file-list>
         <welcome-file>index.jsp</welcome-file>
     </welcome-file-list>
-
 
     <filter>
         <filter-name>struts2</filter-name>
